+++ src/Driver/Sqlite.php
@@ -46,14 +46,13 @@ class Sqlite extends Driver
      * @return string
      * @uses $config
      * @uses getCompression()
-     * @uses getValidCompressions()
      */
     protected function getExportExecutable($filename)
     {
         $compression = $this->getCompression($filename);
         $executable = sprintf('%s %s .dump', $this->getBinary('sqlite3'), $this->config['database']);
 
-        if (in_array($compression, array_filter($this->getValidCompressions()))) {
+        if (in_array($compression, array_filter(VALID_COMPRESSIONS))) {
             $executable .= ' | ' . $this->getBinary($compression);
         }
 
@@ -66,14 +65,13 @@ class Sqlite extends Driver
      * @return string
      * @uses $config
      * @uses getCompression()
-     * @uses getValidCompressions()
      */
     protected function getImportExecutable($filename)
     {
         $compression = $this->getCompression($filename);
         $executable = sprintf('%s %s', $this->getBinary('sqlite3'), $this->config['database']);
 
-        if (in_array($compression, array_filter($this->getValidCompressions()))) {
+        if (in_array($compression, array_filter(VALID_COMPRESSIONS))) {
             $executable = sprintf('%s -dc %s | ', $this->getBinary($compression), $filename) . $executable;
         } else {
             $executable .= ' < ' . $filename;