+++ tests/TestCase/Driver/PostgresTest.php
@@ -131,24 +131,6 @@ class PostgresTest extends DriverTestCase
     }
 
     /**
-     * Test for `getExtension()` method
-     * @test
-     */
-    public function testGetExtension()
-    {
-        $extensions = [
-            'backup.sql' => 'sql',
-            'backup.sql.bz2' => 'sql.bz2',
-            'backup.sql.gz' => 'sql.gz',
-            'text.txt' => null,
-        ];
-
-        foreach ($extensions as $filename => $expectedExtension) {
-            $this->assertEquals($expectedExtension, $this->Postgres->getExtension($filename));
-        }
-    }
-
-    /**
      * Test for `getImportExecutable()` method
      * @test
      */
@@ -166,24 +148,6 @@ class PostgresTest extends DriverTestCase
 
         $expected = $pgRestore . ' -c -e --dbname=' . $dbnameAsString . ' < backup.sql 2>/dev/null';
         $this->assertEquals($expected, $this->invokeMethod($this->Postgres, $method, ['backup.sql']));
-    }
-
-    /**
-     * Test for `getValidExtensions()` method
-     * @test
-     */
-    public function testGetValidExtensions()
-    {
-        $this->assertEquals(['sql.bz2', 'sql.gz', 'sql'], $this->Postgres->getValidExtensions());
-    }
-
-    /**
-     * Test for `getValidCompressions()` method
-     * @test
-     */
-    public function testGetValidCompressions()
-    {
-        $this->assertEquals(['sql.bz2' => 'bzip2', 'sql.gz' => 'gzip', 'sql' => false], $this->Postgres->getValidCompressions());
     }
 
     /**