+++ tests/TestCase/Driver/SqliteTest.php
@@ -123,24 +123,6 @@ class SqliteTest extends DriverTestCase
     }
 
     /**
-     * Test for `getExtension()` method
-     * @test
-     */
-    public function testGetExtension()
-    {
-        $extensions = [
-            'backup.sql' => 'sql',
-            'backup.sql.bz2' => 'sql.bz2',
-            'backup.sql.gz' => 'sql.gz',
-            'text.txt' => null,
-        ];
-
-        foreach ($extensions as $filename => $expectedExtension) {
-            $this->assertEquals($expectedExtension, $this->Sqlite->getExtension($filename));
-        }
-    }
-
-    /**
      * Test for `getImportExecutable()` method
      * @test
      */
@@ -157,24 +139,6 @@ class SqliteTest extends DriverTestCase
 
         $expected = $sqlite3 . ' /tmp/example.sq3 < backup.sql 2>/dev/null';
         $this->assertEquals($expected, $this->invokeMethod($this->Sqlite, $method, ['backup.sql']));
-    }
-
-    /**
-     * Test for `getValidExtensions()` method
-     * @test
-     */
-    public function testGetValidExtensions()
-    {
-        $this->assertEquals(['sql.bz2', 'sql.gz', 'sql'], $this->Sqlite->getValidExtensions());
-    }
-
-    /**
-     * Test for `getValidCompressions()` method
-     * @test
-     */
-    public function testGetValidCompressions()
-    {
-        $this->assertEquals(['sql.bz2' => 'bzip2', 'sql.gz' => 'gzip', 'sql' => false], $this->Sqlite->getValidCompressions());
     }
 
     /**